requirements.txt
Flask[async]==2.3.2
Werkzeug==2.3.6
requests==2.31.0
pandas==2.0.2
numpy==1.24.4
beautifulsoup4==4.12.2
gunicorn==21.2.0
python-dotenv==1.0.0
supabase==2.1.0
redis==5.0.8
nltk==3.8.1
flask-cors==4.0.0
aiohttp==3.9.1
scikit-learn==1.3.2
joblib==1.3.2
bs4==0.0.1
lxml==4.9.3
redis==5.0.8
redis[hiredis]==5.0.8
asyncpg